requirements.txt
numpy
opencv-python
pymavlink
pyserial
PyYAML
#source venv/bin/activate